Worked on category

+++ src/main/java/com/HarvestHopper/service/CategoryServiceImpl.java
@@ -1,4 +1,5 @@
 package com.HarvestHopper.service;
+import com.HarvestHopper.exception.CategoryNotFoundException;
 import com.HarvestHopper.model.Category;
 import com.HarvestHopper.repository.CategoryRepository;
 import org.springframework.stereotype.Service;
@@ -27,6 +28,9 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteCategory(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new CategoryNotFoundException("Category with ID " + id + " not found");
+        }
         categoryRepository.deleteById(id);
     }
